Refactored some classes out of the OsrsDataContainer file. Also updated xpath for image links.
